POST transfer working

+++ app/src/main/java/com/kg/gettransfer/network/Location.kt
@@ -10,7 +10,7 @@ class Location(
         val name: String,
         val point: String) {
 
-    constructor(name: String, la: Int, lo: Int) : this(name, "(" + la.toString() + "," + lo.toString() + ")")
+    constructor(name: String, la: Double, lo: Double) : this(name, "(" + la.toString() + "," + lo.toString() + ")")
 
     fun toMap(): Map<String, String> {
         return mapOf("name" to name, "point" to point)
